+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Release History
 ### Pending
 #### Update
 - chore: throw an exception when the API does not provide streaming support. ([#878](https://github.com/StellarCN/py-stellar-base/pull/878))
+- feat: Add `Asset.contract_id()` for calculating the id of the asset contract. ([#901](https://github.com/StellarCN/py-stellar-base/pull/901))
 
 ### Version 9.2.0